Re-implement Request#pause() to allow partial clearing of requests.
RequestManager#pauseRequests expects to pause requests that are running
without affecting requests that have completed. From a user perspective,
pauseRequests should allow grey squares in a photo grid to remain grey,
without affecting any images that have managed to successfully load.

Without an interface change it's not possible to apply this behavior to
thumbnail requests where the thumbnail(s) have completed but the full
request has not because clear() applies to all parts of the request,
regardless of whether or not they have completed.

The pause() method added in this change allows us to clear() an in
progress full request while leaving any completed thumbnails in place.
As a result the user may see lower resolution thumbnails until
RequestManager#resumeRequests is called, but they won't see the lower
resolution thumbnails disappear as soon as RequestManager#pauseRequests
is called.

PiperOrigin-RevId: 259795516